Minimal version of VP file in HDF format
Save and load the values of variational parameters in HDF format.
Currently only the list of parameter values is saved.
This will work for cases where the wavefunction that saved the values is
structurally the same as the wavefunction that loads the values.
The only validation of this condition is checking the total number of
parameters.
Eventually information about the structure of the wavefunction
will be added to the file in order to interpret the parameters.