Add ev_target parameter for choosing eigenvalue
In theory, the linear method should use the eigenvector corresponding
 to the lowest eigenvalue.

In practice, there are some spurious eigenvalues due to noise.

It appears the QMCPACK code looked for the eigenvalue closest to the existing
energy (H(0,0), called zerozero in the code).

Commit 4b5b0a3f81ef25412eee1350ad579eef62a8f3f3 (Feb 10, 2011) changed this to
look for the eigenvalue closest to 2 Hartrees below the the existing energy (H(0,0) - 2.0)

That commit also changed the window of filter out bad eigenvalues from 1e10 to 100.

It's not known where the magic value of "2.0" comes from?
This commit makes it a parameter (ev_target) that can be adjusted for exploring.